Refactor Timeout into its own module
